Inside a Redux Store
This small version of a Redux store works well enough that you could use it to replace the actual Redux createStore function you've been using in your app so far. (Try it and see for yourself!) The actual Redux store implementation is longer and a bit more complicated, but most of that is comments, warning messages, and handling some edge cases.

As you can see, the actual logic here is fairly short:

The store has the current state value and reducer function inside of itself
getState returns the current state value
subscribe keeps an array of listener callbacks and returns a function to remove the new callback
dispatch calls the reducer, saves the state, and runs the listeners
The store dispatches one action on startup to initialize the reducers with their state
The store API is an object with {dispatch, subscribe, getState} inside
To emphasize one of those in particular: notice that getState just returns whatever the current state value is. That means that by default, nothing prevents you from accidentally mutating the current state value! This code will run without any errors, but it's incorrect: