Middleware
Redux middleware are written as a series of three nested functions. Let's see what that pattern looks like. We'll start by trying to write this middleware using the function keyword, so that it's more clear what's happening:

Whenever an action is dispatched:

The first part of the handleAction function runs, and we print 'dispatching'
We pass the action to the next section, which may be another middleware or the real store.dispatch
Eventually the reducers run and the state is updated, and the next function returns
We can now call storeAPI.getState() and see what the new state is
We finish by returning whatever result value came from the next middleware
Any middleware can return any value, and the return value from the first middleware in the pipeline is actually returned when you call store.dispatch(). For example:

Middleware Use Cases
So, what can we do with middleware? Lots of things!

A middleware can do anything it wants when it sees a dispatched action:

Log something to the console
Set timeouts
Make asynchronous API calls
Modify the action
Pause the action or even stop it entirely
and anything else you can think of.

In particular, middleware are intended to contain logic with side effects. In addition, middleware can modify dispatch to accept things that are not plain action objects. We'll talk more about both of these in Part 6: Async Logic.